Refactor MainController by delegating operations
